maintenances/pyproject.toml
[tool.poetry]
name = "maintenances"
version = "0.1.0"
description = ""
authors = ["josh-work <jdickman@cenic.org>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
keyring = "^23.5.0"
pyotp = "^2.6.0"
napalm = "^3.3.1"
atlassian-python-api = "^3.18.1"
textfsm = "^1.1.2"
lxml = "^4.7.1"
netaddr = "^0.8.0"
pandas = "^1.4.0"
pygsheets = "^2.0.5"
pydantic = "^1.9.0"
typer = "^0.4.0"
ipaddress = "^1.0.23"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
